move EVS application json configs into its folder
These is just an auxiliary change that intended to make dependencies
more convenient.

The config json files are used by the EVS application at run-time. They
are stored in the EVS HAL folder due to historical reasons. When the
AOSP version of the EVS application used to use it was a proper way to
have platform-specific configuration files. But now when we have our own
fork of the EVS application it would be more correct to store them in
the EVS application project folder.

Note: this changes must be applied simultaneously for the following
three projects:
 - $ANDROID_BUILD_TOP/vendor/renesas/hal/evs;
 - $ANDROID_BUILD_TOP/vendor/renesas/apps/evs;
 - $ANDROID_BUILD_TOP/device/renesas/common.

Test: build, flash a dut and ensure that EVS works well.
Jira-Id: ADM-5092
Change-Id: I835e8a57226ad2f1bc7d657b25b0446b87698642
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>